\name{DFA.CANCOR-package}

\alias{DFA.CANCOR-package}

\docType{package}

\title{DFA.CANCOR}

\description{This package provides SPSS- and SAS-like output for linear discriminant 
	function analysis (via the DFA function) and for canonical correlation analysis
	(via the CANCOR function).}

