\name{DFA.CANCOR-package}

\alias{DFA.CANCOR-package}

\docType{package}

\title{DFA.CANCOR}

\description{Provides SPSS- and SAS-like output for linear discriminant 
	function analysis (via the DFA function) and for canonical correlation analysis
	(via the CANCOR function). There are also functions for assessing the assumptions
	of normality, linearity, and homogeneity of variances and covariances.}

